backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0
pydantic[email]==2.5.0
python-dotenv==1.0.0
fastapi-mail==1.4.1
reportlab==4.0.7
openpyxl==3.1.2
pandas==2.1.4
jinja2==3.1.2
aiofiles==23.2.1
pillow==10.1.0
qrcode==7.4.2
twilio==8.10.0
celery==5.3.4
redis==5.0.1